docs/requirements.txt
jinja2
mkdocs-autorefs
mkdocs-awesome-nav
mkdocs-macros-plugin
mkdocs-material>=9.0
pyyaml
